Combine iBase and iGeom.
